pyproject.toml
[tool.poetry]
name = "mqtt-client-dht22"
version = "0.1.0"
description = ""
authors = ["Matheus Mendez <matheusmendez88@gmail.com>"]
readme = "README.md"
packages = [{include = "mqtt_client_dht22"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
mysql-connector-python = "^8.0.32"
numpy = "^1.24.2"
datetime = "^5.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.10.0"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
